+++ src/typed_settings/_file_utils.py
@@ -1,5 +1,5 @@
 from pathlib import Path
-from typing import Iterable, Union
+from typing import Iterable, Optional, Union
 
 
 ROOT_DIR = Path().resolve().root
@@ -9,6 +9,8 @@ def find(
     filename: str,
     stop_dir: Union[str, Path] = ROOT_DIR,
     stop_files: Iterable[str] = (".git", ".hg"),
+    *,
+    start_dir: Optional[Path] = None,
 ) -> Path:
     """
     Search for a file in the current directory and its parents and return its
@@ -22,6 +24,7 @@ def find(
         stop_dir: Stop searching if the current search dir equals this one.
         stop_files: Stop searching if the current search dir contains
          this file.
+        start_dir: Start directory for the search, defaults to the current directory.
 
     Returns:
         The resolved path to *filename* if found, else ``Path(filename)``.
@@ -36,8 +39,12 @@ def find(
 
             find("pyproject.toml", stop_file=".git")
 
+    .. versionchanged:: 23.1.0
+       Added the *start_dir* parameter.
     """
-    start = Path.cwd().joinpath(filename)
+    if start_dir is None:
+        start_dir = Path.cwd()
+    start = start_dir.joinpath(filename)
     for path in start.parents:
         p = path.joinpath(filename)
         if p.exists():